backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
playwright==1.47.0
python-socketio==5.11.0
celery[redis]==5.4.0
redis==5.1.0
openai==1.50.0
reportlab==4.2.0
python-pptx==1.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
cssutils==2.11.0
esprima==4.0.1
aiofiles==24.1.0
apscheduler==3.10.4
